Moved rotation code into test
