+++ case_studies/fnp/one_dim.py
@@ -44,17 +44,13 @@ if path not in sys.path:
     sys.path.insert(0, path)
 import numpy as np
 from bliss.models.fnp import (
-    RegressionFNP,
-    PoolingFNP,
-    OneDimDataset,
-)
-from bliss.models.fnp import (
     FNP,
     SequentialVarg,
     MLP,
     SplitLayer,
     NormalEncoder,
     AveragePooler,
+    SetPooler,
     RepEncoder,
     ConcatLayer,
     DepGraph,
@@ -80,10 +76,7 @@ def train_onedim_model(model, od, epochs=10000, lr=1e-4, visualize=False):
     holdout_loss_initial = model(od.XR, od.yhR, od.XM, od.yhM)
     holdout_loss_best = holdout_loss_initial
     print("Initial holdout loss: {:.3f})".format(holdout_loss_initial.item()))
-    if isinstance(model, RegressionFNP):
-        stdy = None
-    else:
-        stdy = od.stdy
+    stdy = od.stdy
     for i in range(epochs):
         optimizer.zero_grad()
 
@@ -221,10 +214,172 @@ def visualize_onedim(
 
 
 # %%
+class OneDimDataset:
+    def __init__(
+        self,
+        N=20,
+        num_extra=500,
+        seed=1,
+        offset=0.1,
+    ):
+        ## Generate the first row as in the FNP paper
+        np.random.seed(seed)
+        X = np.concatenate(
+            [
+                np.random.uniform(low=0, high=0.6, size=(N - 8, 1)),
+                np.random.uniform(low=0.8, high=1.0, size=(8, 1)),
+            ],
+            axis=0,
+        )
+        eps = np.random.normal(0.0, 0.03, size=(X.shape[0], 1))
+        self.f = lambda x, eps: x + np.sin(4 * (x + eps)) + np.sin(13 * (x + eps)) + eps
+        y = self.f(X, eps)
+
+        ## Pick which indices are references or not
+        # self.idxR = idxR
+        # self.idxM = np.array([i for i in idx if i not in idxR.tolist()])
+
+        ## Generate more y-values
+        ys = [y]
+        for _ in range(99):
+            Xi = X + np.random.normal()
+            eps_i = np.random.normal(0.0, 0.03, size=(X.shape[0], 1))
+            # f = lambda x, eps: x + np.sin(4 * (x + eps)) + np.sin(13 * (x + eps)) + eps
+            yi = self.f(Xi, eps_i)
+            ys.append(yi)
+        y = np.concatenate(ys, axis=1).transpose()
+
+        ## Generate holdouts
+        ys = []
+        for i in range(10):
+            Xi = X + np.random.normal()
+            eps_i = np.random.normal(0.0, 0.03, size=(X.shape[0], 1))
+            # f = lambda x, eps: x + np.sin(4 * (x + eps)) + np.sin(13 * (x + eps)) + eps
+            yi = self.f(Xi, eps_i)
+            ys.append(yi)
+        yh = np.concatenate(ys, axis=1).transpose()
+
+        self.stdx, self.stdy = StandardScaler().fit(X), StandardScaler().fit(
+            y.reshape(-1, 1)
+        )
+        # X, y = stdx.transform(X), stdy.transform(y)
+        X = self.stdx.transform(X)
+        idx = np.arange(X.shape[0])
+        # self.idxR = np.random.choice(idx, size=(10,), replace=False)
+        # self.idxM = np.array([i for i in idx if i not in idxR.tolist()])
+        self.idxR = np.array([2, 16, 9, 6, 17, 12, 4, 15, 1, 14])
+        self.idxM = np.array([i for i in idx if i not in self.idxR.tolist()])
+
+        self.X = torch.from_numpy(X.astype(np.float32))
+        self.y = torch.from_numpy(y.astype(np.float32)).unsqueeze(2)
+        self.XR, self.yR = self.X[self.idxR], self.y[:, self.idxR]
+        self.XM, self.yM = self.X[self.idxM], self.y[:, self.idxM]
+
+        ## Holdouts
+        yh = torch.from_numpy(yh.astype(np.float32))
+        self.yh = yh.unsqueeze(2)
+        self.yhR = self.yh[:, self.idxR]
+        self.yhM = self.yh[:, self.idxM]
+
+        ## Point where predictions will be made for plotting
+        self.dx = np.linspace(-1.0, 2.0, num_extra).astype(np.float32)[:, np.newaxis]
+
+    def cuda(self):
+        for nm in ["XR", "XM", "X", "yR", "yM", "y", "yhR", "yhM", "yh"]:
+            setattr(self, nm, getattr(self, nm).cuda())
+
+    def cpu(self):
+        for nm in ["XR", "XM", "X", "yR", "yM", "y", "yhR", "yhM", "yh"]:
+            setattr(self, nm, getattr(self, nm).cpu())
+
+        # return x, y, dx, f
+
+
+# %%
 od = OneDimDataset()
 # %%
 if torch.cuda.is_available():
     od.cuda()
+
+# %%
+def make_onedim_model(
+    dim_x=1,
+    dim_y=1,
+    dim_z=50,
+    dim_u=3,
+    dim_y_enc=100,
+    fb_z=1.0,
+    use_plus=False,
+    use_set_pooler=False,
+    pooling_rep_size=32,
+    pooling_layers=[64],
+    use_attention=False,
+    st_numheads=[2, 2],
+):
+    cov_vencoder = SequentialVarg(
+        MLP(dim_x, [100], 2 * dim_u),
+        SplitLayer(dim_u, -1),
+        NormalEncoder(),
+    )
+    dep_graph = DepGraph(dim_u)
+    trans_cond_y = MLP(dim_y, [128], dim_y_enc)
+    if use_set_pooler:
+        rep_encoder = RepEncoder(
+            MLP(dim_y_enc + dim_u, [128], pooling_rep_size),
+            use_u_diff=True,
+            use_x=False,
+        )
+        pooler = SequentialVarg(
+            SetPooler(
+                pooling_rep_size,
+                dim_z,
+                pooling_layers,
+                use_attention,
+                st_numheads,
+            ),
+            SplitLayer(dim_z, -1),
+            NormalEncoder(minscale=1e-8),
+        )
+    else:
+        rep_encoder = RepEncoder(
+            MLP(dim_y_enc + dim_x, [128], 2 * dim_z), use_u_diff=False, use_x=True
+        )
+        pooler = SequentialVarg(
+            AveragePooler(dim_z),
+            SplitLayer(dim_z, -1),
+            NormalEncoder(minscale=1e-8),
+        )
+    prop_vencoder = SequentialVarg(
+        ConcatLayer([1, 0]),
+        MLP(
+            dim_y_enc + dim_x,
+            [128],
+            2 * dim_z,
+        ),
+        SplitLayer(dim_z, -1),
+        NormalEncoder(minscale=1e-8),
+    )
+    output_in = [0] if not use_plus else [0, 1]
+    output_insize = dim_z if not use_plus else dim_z + dim_u
+    label_vdecoder = SequentialVarg(
+        ConcatLayer(output_in),
+        MLP(output_insize, [128], 2 * dim_y),
+        SplitLayer(dim_y, -1),
+        NormalEncoder(minscale=0.1),
+    )
+    model = FNP(
+        cov_vencoder,
+        dep_graph,
+        trans_cond_y,
+        rep_encoder,
+        pooler,
+        prop_vencoder,
+        label_vdecoder,
+        fb_z=fb_z,
+    )
+
+    return model
+
 
 # %% [markdown]
 # ## FNP
@@ -250,40 +405,7 @@ dim_y = 1
 dim_z = 50
 dim_u = 3
 dim_y_enc = 100
-fnp = FNP(
-    cov_vencoder=SequentialVarg(
-        MLP(dim_x, [100], 2 * dim_u),
-        SplitLayer(dim_u, -1),
-        NormalEncoder(),
-    ),
-    dep_graph=DepGraph(dim_u),
-    trans_cond_y=MLP(dim_y, [128], dim_y_enc),
-    rep_encoder=RepEncoder(
-        MLP(dim_y_enc + dim_x, [128], 2 * dim_z), use_u_diff=False, use_x=True
-    ),
-    pooler=SequentialVarg(
-        AveragePooler(dim_z),
-        SplitLayer(dim_z, -1),
-        NormalEncoder(minscale=1e-8),
-    ),
-    prop_vencoder=SequentialVarg(
-        ConcatLayer([1, 0]),
-        MLP(
-            dim_y_enc + dim_x,
-            [128],
-            2 * dim_z,
-        ),
-        SplitLayer(dim_z, -1),
-        NormalEncoder(minscale=1e-8),
-    ),
-    label_vdecoder=SequentialVarg(
-        ConcatLayer([0]),
-        MLP(dim_z, [128], 2 * dim_y),
-        SplitLayer(dim_y, -1),
-        NormalEncoder(minscale=0.1),
-    ),
-    fb_z=1.0,
-)
+fnp = make_onedim_model()
 if torch.cuda.is_available():
     fnp = fnp.cuda()
 
@@ -295,17 +417,7 @@ fnp, _, _, fnp_loss_best = train_onedim_model(
 # ## FNP+
 
 # %%
-fnpp = RegressionFNP(
-    dim_x=1,
-    dim_y=1,
-    transf_y=od.stdy,
-    dim_h=100,
-    dim_u=3,
-    n_layers=1,
-    dim_z=50,
-    fb_z=1.0,
-    use_plus=True,
-)
+fnpp = make_onedim_model(use_plus=True)
 fnpp, _, fnpp_loss_final = train_onedim_model(
     fnpp, od, epochs=10000, lr=1e-4, visualize=VISUALIZE
 )
@@ -314,19 +426,7 @@ fnpp, _, fnpp_loss_final = train_onedim_model(
 # ## Attention FNP
 
 # %%
-attt = PoolingFNP(
-    dim_x=1,
-    dim_y=1,
-    transf_y=None,
-    dim_h=100,
-    dim_u=3,
-    n_layers=1,
-    dim_z=50,
-    fb_z=1.0,
-    use_plus=False,
-    use_direction_mu_nu=True,
-    set_transformer=True,
-)
+attt = make_onedim_model(use_set_pooler=True, use_attention=True)
 attt, _, attt_loss_final = train_onedim_model(
     attt, od, epochs=10000, lr=1e-4, visualize=VISUALIZE
 )
@@ -335,19 +435,7 @@ attt, _, attt_loss_final = train_onedim_model(
 # ## Deep-set FNP
 
 # %%
-poolnp = PoolingFNP(
-    dim_x=1,
-    dim_y=1,
-    transf_y=od.stdy,
-    dim_h=100,
-    dim_u=3,
-    n_layers=1,
-    dim_z=50,
-    fb_z=0.5,
-    use_plus=False,
-    use_direction_mu_nu=True,
-    set_transformer=False,
-)
+poolnp = make_onedim_model(use_set_pooler=False, use_attention=False, fb_z=0.5)
 poolnp, _, poolnp_loss_final = train_onedim_model(
     poolnp, od, epochs=10000, lr=1e-4, visualize=VISUALIZE
 )
